extract as function: `assert_primed_adjacent`
because it is not a hard constraint. Currently, it applies to `dd.bdd.rename`. It does not apply when using `cudd.rename`. So `assert_primed_adjacent` should be called explicitly, only when needed. This avoids over-constraining `_extract_partition`.

Nonetheless, swapping with non-adjacent variables can be performed by reordering to bring the pairs together, use `rename`, then reorder back. This may be implemented in the future. It is encouraged though to group together each pair of primed and unprimed variables.