add `Automaton` arg: `add`, to create any missing variables
this allows passing an existing `BDD`, and augmenting its variables. So it allows bitblasting extra variables, with the same `BDD` as target. This is convenient when using `dd.cudd`, because it does not require to re-create a large `BDD`. For example, one can add counters of winning goals, for constructing transducers.