add long subformulae first to `BDD`
because it has experimentally been observed to have a significant effect. The heuristic justification is that, initially, the BDD variable order is easier to affect. So, adding the large formulae first tends to affect the order towards serving those formulae.

In contrast, adding a small formula first would change the order to efficiently express that formula. Adding a larger formula later, it will be more difficult to obtain a good order for compressing that second formula, because the order is already constrained by the existing nodes (created for the small formula).