BUG: prime integer limits
because this allows the environment to choose a next value that violates the limits, provided that sys will lose in the next time step. The reason is that this value would force env to lose later, and by that time env has won (because, by that time, sys has violated its action).

An initial condition needs to be added, so that it can be copied when
adding saturating counters to a strategy. Otherwise, closure is not
taken for the transducer memory (counters), so the initial condition for
those variables is not imposed by the winning set (and would be present only
in the action, not in the initial condition).

As previously, constrain unprimed integer values, because otherwise
invalid states are possible as starting states, which relies on correct handling of the initial condition to avoid errors. Thus, it is susceptible to errors with the initial conditions code/specification, so not as robust (explicit better than implicit). It also reduces the number of possible predecessor states during fix point computations, and simplifies sets during debugging.