use "<>[]" and "[]<>" as winning condition keys
because otherwise uniform treatment of (generalized) Rabin and Streett games becomes very confusing.

Likewise, "env" and "sys" correspond to universal and existential quantification, and not to the players. In other words, computing a counterstrategy requires swapping and negating recurrence and persistence goals, as well as initial conditions and actions.

For this reason, the keys "env" and "sys" may be replaced in the future.