API: add missing bits in all cases in `_bitvector_to_bdd`
To require that bits should already be added is cumbersome,
because the user would need to create and add primed bits,
and then reorder to the desired order.

The addition can be done automatically,
and the only issue is obtaining a desired bit order.
Define levels only if no bits already exist in manager (simpler).
Otherwise, this is a different task and should be done separately (flat).

No extra function for adding variables.
No `bdd` arg, because that can be set outside (flat).
The case `add=False` was not used anywhere.

From the viewpoint of `Automaton.build`,
A BDD manager loaded from file and a manager already initialized
are indistinguishable.
(for example, by copying bits from another manager)
Both cases are more specialized,
so call directly internal functions as needed.

For different needs, override `build`, or
write your own function for building the BDDs.