API: extract function `type_invariants` from `bitblast_table`
and do not use information about players, so that bitblasting be decoupled from variable ownership.

Type information is needed to bitblast. But the type information can originate from:

1. type declarations in a typed language (e.g., Promela), in which case `type_invariants` would need to be called after `bitblast_table`, or

2. a type invariant within a TLA+ specification, in which case `type_invariants` need not be called.

Besides, `bitblast_table` can be called also for other purposes of using BDDs, when the initial condition and type invariant are not needed.

Also, returning a symbol table (instead of a `tuple`) from `bitblast_table` is less confusing, especially for use within an interpreter.