API: do not allow "previous" operator in arithmetic context
This experiment would never work simply enough.
Also, it would require propagating type information upwards, depending on the bit width of arithmetic operations, so that auxiliary integer variables be created with a suitable range.

Remove the confusing delayed lookup of type information. Instead, just propagate the symbol table to the flattener, and check that only Boolean variables appear as operands of "previous" operators.