MAI: allow 1 with `@` operator in bitblasting syntax
assertion introduced in 4470bf08bfb0ad3516d20175900e46c780919d2a.
The assertion guarded early for the error of passing 0 or 1 to
`dd.cudd.BDD`, which was the original use case for `@`.

`dd.cudd.BDD` creates integer references from C pointer values [1].
Therefore, the values 0 and 1 are *unrelated* to the truth
values FALSE and TRUE. This is already guarded for in the methods
`dd.cudd.BDD._add_int` and `dd.cudd.Function.__int__`.

`dd.bdd.BDD` returns the nodes themselves as their integer
syntactic form for `@` mentions [2]. It so happens that node 1
*is* the truth value TRUE. So, an additive shift is unnecessary
for `dd.bdd.BDD`. Thus, "@ 1" should be allowed in bitblasting.

"@ 0" still raises an `AssertionError`. Neither `dd.cudd.BDD` nor
`dd.bdd.BDD` use 0 for integer `@` mentions.

It so happened that code using expressions containing `@` was
demanding enough to require reordering, so I run it with
`dd.cudd`, instead of `dd.bdd`. So, this restriction
remained in place until now.

Dynamic reordering has been implemented for `dd.bdd` and
is forthcoming in `dd == 0.5.0`. So, this restriction is
over-constraining.

[1] https://github.com/johnyf/dd/commit/e7fd76d41f40fbdb8de51d67fe724058cca7d628
[2] https://github.com/johnyf/dd/commit/5b8ca75f1dd915b2ca67cc9ffe4c4e817056aaf1