API: use `dd.autoref` if `fol` fails to import `dd.cudd`
instead of `dd.bdd`. The reason is a shift to a shift to the
high-level API that both `dd.cudd` and `dd.autoref` use.
This ensures automated reference management, which avoids
errors and utilizes Python's garbage collector via `weakeref`.

Also, dynamic reordering in `dd.bdd` as of `dd == 0.5.0`
necessitates `incref` in user code. A tedious requirement
unless automated reference management is used.

When code written with `dd.cudd` underneath is suddenly run
with the low level `dd.bdd`, it can fail in subtle ways,
which isn't the case with `dd.autoref`, due to the common
interface with `dd.cudd`.