BUG: exhaust the enumeration in `fol.Context.pick_iter`
Only one value of the generator was consumed, instead of
the entire enumeration. This affected only calls without
`care_vars`, which anyway raised an `AssertionError` in
`dd < 0.5.0`, because `care_vars` was replaced with an empty
set, and `BDD.sat_iter` interpreted `care_bits` as a constraint.

In other words, this error in enumeration matters only for
bits that aren't enumerated exhaustively by `sat_iter`.
A call with `care_vars` had to include the entire support,
and each integer results in all its bits being included in
`care_bits`, so there weren't any bits not exhaustively enumerated.

Trying to omit integers in support from `care_vars` resulted
in errors in `sat_iter`. This behavior will change in
`omega == 0.1.0`, so this incorrect enumeration would
become observable, if left uncorrected.