TST: rm `del` statements for BDD nodes
because `dd >= 0.5.0` avoids circular references between
nodes and manager to utilize Python's garbage collector in
`autoref` and `cudd`. This ensures that nodes are collected
before the manager. It prevents the opposite behavior.

For example, `dd.cudd.Function.bdd` is a reference to the manager.
The manager used to reference an instance of `Function` for the
constant nodes, but not any more. So, the references form an
acyclic graph.