ENH: parse operator definitions with `logic.lexyacc.Parser`
Same parser behavior as for expressions. A consecution of operator
definitions is parsed as a list of operator definitions.

Store the operator name in a `Terminal` to ensure that no
errors are raised when `str` is called on an AST.
Otherwise, we would get the error:

"TypeError: wrapper __str__ doesn't take keyword arguments"