ENH: add functions to analyze BDD support to module `omega.symbolic.bdd`
functions added:

- `joint_support`
- `support_issubset`
- `is_state_predicate`
- `is_primed_state_predicate`
- `is_proper_action`
- `print_support`
- `prime`
- `unprime`

Prefer the function `support_issubset` to a function `assert_support`
for the following reasons:

- Allows using the result differently than for raising errors.

- Allows using `raise AssertionError` to ensure that optimizations
  won't prevent the error from being raised.

- Allows using `raise SomeOtherException`

- Allows using the returned value to control execution,
  instead of raising an error.

- The user has to write a little more. Nonetheless, the `assert`
  statement is highlighted in user code, thus more readable and
  explicit, and also counted towards the assertion density.

- `assert support_issubset` is more directly understood than
  `assert_support`. The latter requires looking at the docstring of
  the function the first time one encounters it.

- "issubset` is less readable, but more explicit.

We could write these functions with `u` as the only argument,
by using the property `dd.cudd.Function.support` (`dd == 0.5.0`).
However, that would show us bits, not `fol.Context.vars`.

The functions `is_state_predicate, is_primed_state_predicate,
is_proper_action` are all checking whether variables are primed.
The bits that represent a primed integer-valued variable are
primed too, so these functions can work with the bit-support.
So we don't need to pass `fol` as argument.

This approach is also faster, because the bit support is not
converted to a support that contains integer-valued variables.