BUG: declare temporal `Phi` and action `e` as operator arguments
TLA+2 allows STATE, TEMPORAL to appear only within an ASSUME / PROVE [1].
So the temporal property `Phi` needs to be passed as an argument.
Passing `e` as an argument is more convenient too.

`phi` was declared as a `CONSTANT` to enable checking with SANY,
with `TEMPORAL` within a comment.
`mu` has been renamed to `e`.

[1] Leslie Lamport, "TLA+2: A preliminary guide", 2014
    http://lamport.azurewebsites.net/tla/tla2-guide.pdf