MAI: rm call to `symbolic._prime_and_order_table`
This function is obsolete, in two ways:


1. In the old class `symbolic.Automaton` only unprimed first-order
variables are delcared, and string init, action, win.
Primed identifiers are added only as bits at the time of
building a BDDized automaton. The function `symbolic._add_primed_bits`
addes them, not `_prime_and_order_table`.

Priming identifiers and at the same time ordering BDD levels
mixes two separate concerns. This is why these operations
were separated at some point, by transitioning to calling
the functions `symbolic._add_primed_bits` and
`symbolic._pick_var_order`.

Picking an order of BDD levels is an optimization, and
unrelated to declaring variables and encoding them as bitvectors.
So `_pick_var_order` shouldn't be part of `Automaton`
internals. This is the case for the new class `temporal.Automaton`.

It is convenient to prime first-order variables, so that we
can work with predicates, support, etc. that involve primed
variables at the first-order level. This is supported by
`temporal.Automaton`, due to how `fol.Context` is designed.

In conclusion, the function `_prime_and_order_table` is
unused in `symbolic.Automaton`, and it should be avoided.


2. No `fol.Context` is created any more in `games.enumeration`.
So there is no need to initialize variables in any way.
Even if there was, `_prime_and_order_table` is unnecessary,
because the method `add_variables` of `temporal.Automaton`
provides priming functionality of first-order variables.