REF: use `aut.varlist`, instead of creating `aut.control`
The new class `temporal.Automaton` contains the partition
of variables already within the attribute `aut.varlist`,
which is a `dict` that maps each player name to a `list`
of variable names.

Return primed system and environment variables from the
function `_split_vars_per_quantifier`. The method
`temporal.Automaton.prime_vars` could be used, but it
introduces an assumption about API, which is easy to
prevent with a list comprehension.

Another alternative would be to access lists of primed
variables `aut.varlist["env'"]` and `aut.varlist["sys'"]`.
Likewise, doing so introduces an API assumption.