MAI: define `keys` using only unprimed identifiers
So constants and unprimed variables.
The attribute `symbolic.Automaton.vars` includes only unprimed
variables, but in the replacement class, the attribute
`temporal.Automaton.vars` includes also primed identifiers
of variables.

Setting `varnames` to all the variables may be too restrictive,
because of the assertion that the node annotation assigns
values to all variables. This annotation results by combining
an assignment to variables in `aut.varlist['env']` and
an assignment to variables in `aut.varlist['sys']`.
If more variables are declared in `aut.vars` than those of
env and sys, then the assertion will fail.

In the future, the assertion `set(values) == varnames`
may be relaxed to `set.issubset`.