DEP: `omega.symbolic.symbolic.Automaton`
Regarding use of `PendingDeprecationWarning` and
`DeprecationWarning`, see [PEP 387]. Quoting:

"Add a warning [2]. If behavior is changing, the API may
gain a new function or method to perform the new behavior;
old usage should raise the warning. If an API is being
removed, simply warn whenever it is entered.

DeprecationWarning is the usual warning category to use,
but PendingDeprecationWarning may be used in special cases
were the old and new versions of the API will coexist for
many releases."