API: add positional args `env` and `sys` to `action_to_steps`
explicit is better than implicit [PEP 20].
Synthesis creates the initial condition and action of
a component named "impl", by adding the key `'impl'` to
the attribute `action` and `'impl_env'` and `'impl_sys'`
to `init` (the latter is a peculiarity, which will be
tidied soon).

Passing the automaton to `action_to_steps` after synthesis,
while forgetting to copy `action['impl']` to `action['sys']`
and `init['impl_env']` to `init['env']` will silently
lead to enumeration of the wrong thing.

Moreover, the user shouldn't be forced to modify the "env"
and "sys" keys. The simplest way for avoiding this error
is to require from the user to explicitly name which
component to enumerate, hence this commit.

After forthcoming API changes to handling of initial
conditions, the synthesis result will be stored as
a single component named "impl". The environment will be
common for the implementation and specification.
So the user will be able to pass `env='env', sys='impl'`,
without need to copy anything.