BUG: append sign bit to
When a sign bit is present, it is correctly taken into account.
However, when a sign bit is absent, and the type hint ranges over
only negative values, the bitvector values are interpreted as positive.
They should be negated, in order to return negative values.

Introduced in 5dad03412fc7bbbc7443125d4c75d3a316830a81.