BUG: check the expression returned by `Context.to_expr`
The function `list_expr` was called with `use_dom=False`.
In principle, this is correct. However, it could result in
a different DNF being checked in `to_expr` than the
DNF returned by the function `dumps_cover`.

The reason is that the function `dumps_cover` modifies the
value of `show_dom`. To ensure that the same value of
`show_dom` is used in both calls to `list_expr`, move the
assertion to inside `dumps_cover`. Checking there serves also
other calls to `dumps_cover`.

The DNF checked is still slightly different because of calling
`stx.disj` instead of `stx.vertical_op`.
It differs also in case `latex=True`.

Remove the call of `dumps_cover` from `cyclic_core`, because
it passes essential elements only, so the new assertion can fail.