API: utilize arg `vrs` in `implies_type_hints`
The argument `vrs` was unused. This change makes `vrs`
an optional argument. So `u` can be required to imply
the type hints that correspond to a specific set of variables,
for example the support of `u`.

Previously, calls to `implies_type_hints` would typically
return `False`, because most BDDs depend on only a subset
of the declared variables.