API: change arg `aut` type to `temporal.Automaton`
- change type of argument `aut` to `temporal.Automaton`
  for the following functions in `omega.symbolic.enumeration`:
  - `dump_relation()`
  - `relation_to_graph()`
  - `print_edges()`
  - `_make_table()`

- rm argument `prime` from function `_care_relation()`

- add the function `omega.symbolic.enumeration._prime_bits()`
  to prime the nonconstant bits in a BDD `Function`

- update the test function `test_relation_to_graph()` in
  the script `symbolic_enumeration_test.py`

More detailed comments about the changes in the module
`omega.symbolic.enumeration` follow.


## Variable `keys`

The variable `keys` in the function `relation_to_graph()`
used to be defined by enumerating `aut.vars`.
The attribute `omega.symbolic.symbolic.Automaton.vars`
contained only unprimed variables (by the bitblasting that
takes place in the method
`omega.symbolic.symbolic.Automaton.build()`).

The attribute `omega.symbolic.temporal.Automaton.vars`
contains both unprimed and primed identifiers. This is why
the variable `keys` is now defined by enumerating only the
unprimed identifiers in `aut.vars`.


## Constants unsupported

The unprimed identifiers in `aut.vars` can now include also
constants. There were no constants in `symbolic.Automaton`.

Identifiers in a model are split between the `source` and `target`
models later within the function `relation_to_graph()`.
Unprimed identifiers are placed in `source`, and primed
identifiers in `target`. So `target` cannot contain any
keys that are constants.

The primed identifiers in `target` are unprimed (by calling
the function `_unprime_any_primed()`), so `target` at this
point cannot have any constants as keys.

Then it is required inside the function `_find_or_add_model()`
that all elements of `keys` be keys of the model `target`.
Therefore, if any constants are present in `aut.vars`,
then those will be in `keys`, and a `KeyError` will be raised
inside the function `_find_or_add_model()`.

This reasoning implies that automata with constants are
not yet supported within the functions of the module
`omega.symbolic.enumeration` that take `aut` as an argument.
Constants were not supported previously by these functions,
because the attribute `omega.symbolic.symbolic.Automaton.vars`
does not contain any constants.


## Splitting between `source` and `target` based on priming

The splitting between `source` and `target` used to be done
depending on whether `var in aut.vars`. Since `aut.vars` used
to contain only unprimed variables, this is equivalent to
splitting based on whether an identifier is primed.
This is done now.

The only difference is that now an unprimed identifier is not
necessarily an unprimed variable, but can be a constant.
As discussed above, constants are not yet supported in the
module `enumeration`, so this case is outside the
supported argument values.


## Outline to support constants

In the future, the splitting between `source` and `target`
could be adapted to support constants. The adaptation
could be to split identifiers as currently done, then
unprime the primed variables in `target`, and then
copy to `target` only those items of `source` that have
constants as keys.

This approach correctly represents that the values of
constants remain unchanged through behavior steps.
Moreover, with this approach no `KeyError` will be
raised inside the function `_find_or_add_model()`,
because `target` will have all elements of `keys`
as keys (i.e., both unprimed variables and constants).


## Removing the argument `prime` from function `_care_relation()`

The removal of the argument `prime` from the function
`_care_relation()` was necessary because the class
`omega.symbolic.temporal.Automaton` does not have an
attribute `prime` (the class `omega.symbolic.symbolic.Automaton`
has such an attribute).

Previously, the attribute `aut.prime` was used in the
function `_care_relation()` to rename the bits in the
support of the BDD `target` (by priming those bits).

The attribute `aut.prime` was defined in the function
`omega.symbolic.symbolic._bitvector_to_bdd()` to be
a dictionary that maps unprimed to primed bits, only for
those bits that correspond to variables defined as
player variables.

Now the function `_prime_bits()` primes all bits in the
support of `target` that correspond to variables
(nonconstants). So any non-player variables in the
support of `target` will now be primed.

It seems to make sense to take all variables into account
when priming (i.e., not only those that are player
variables). The previous design may have been so as
a side effect of design decisions outside of the module
`omega.symbolic.enumeration`.

(Constants were not contained in `aut.prime`,
so they could not be contained in the support of
the BDD `target`. Now the function `_prime_bits()`
primes only those bits that correspond to nonconstants.
This allows constants to appear in the support of
`target`.)


## Copying `aut.vars` instead of calling `_prime_and_order_table()`

The attribute `aut.vars` is deep-copied now to define variable
`t` within the function `_make_table()`. Previously the
function `omega.symbolic.symbolic._prime_and_order_table()` was
called to define `t` using `aut.vars`.

As already mentioned, the attribute `aut.vars` now contains
both unprimed identifiers (unprimed variables and constants),
as well as primed identifiers (primed variables).

Previously, `aut.vars` contained only unprimed variables,
but the function `_prime_and_order_table()` added primed
variables. So the resulting `dict` `t` has the same keys
(except that also constants can be present now). The only
difference is that the value of each key in `t` used to
have also the keys `'level'` and `'len'` (which were created
by the function `_prime_and_order_table()`).

But the keys `'level'` and `'len'` are unused in the
function `_bitfields_to_int_iter()`, which is where `t`
propagates via the functions `_enumerate_bdd()` and
`_print_enumeration()`, after `t` is defined from the
value returned by the function `_make_table()`.
So there is no need to create these keys.

This observation is implied also by noting that
`_bitfields_to_int_iter()` is called also
from the method `omega.symbolic.fol.Context.pick_iter()`
with the data attribute `omega.symbolic.fol.Context.vars`
as argument.

This is essentially the same attribute as now given as
argument to `_bitfields_to_int_iter()` from
`_enumerate_bdd()`. The reason is that the class
`omega.symbolic.temporal.Automaton` subclasses the
class `omega.symbolic.fol.Context`.